login reducer to profile reducer
